requirements.txt
celery==5.6.3
channels==4.3.2
channels-redis==4.3.0
coverage==7.13.5
django==6.0.3
django-admin-thumbnails==1.0.0
django-ckeditor-5==0.2.20
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-paypal==2.1
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
djlint==1.36.4
djoser==2.3.3
flake8==7.3.0
gunicorn==25.3.0
httpx==0.28.1
openai==2.30.0
pillow==12.1.1
prometheus-client==0.24.1
psycopg2==2.9.11
redis==7.4.0
stripe==11.5.0
uvicorn==0.42.0
websockets==16.0
